Update src/AddressBookProject.java
fixed input bugs
edited ifEmpty evaluation
formatted code

+++ src/AddressBookProject.java
@@ -1,110 +1,130 @@
 import java.util.ArrayList;
 import java.util.Scanner;
 
-class Contact{
+class Contact {
     private String firstName;
     private String lastName;
     private String email;
     private String phone;
-    public Contact(String firstName, String lastName, String email, String phone){
-        this.firstName=firstName;
-        this.lastName=lastName;
-        this.email=email;
-        this.phone=phone;
+
+    public Contact(String firstName, String lastName, String email, String phone) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.phone = phone;
     }
-    public void setFirstName(String firstName){
-        this.firstName=firstName;
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
     }
-    public String getFirstName(){
+
+    public String getFirstName() {
         return firstName;
     }
-    public void setLastName(String lastName){
-        this.lastName=lastName;
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
     }
-    public String getLastName(){
-        return  lastName;
+
+    public String getLastName() {
+        return lastName;
     }
-    public void setEmail(String email){
-        this.email=email;
+
+    public void setEmail(String email) {
+        this.email = email;
     }
-    public String getEmail(){
+
+    public String getEmail() {
         return email;
     }
-    public void setPhone(String phone){
-        this.phone=phone;
+
+    public void setPhone(String phone) {
+        this.phone = phone;
     }
-    public String getPhone(){
+
+    public String getPhone() {
         return phone;
     }
-    public void printContact(){
-        System.out.println("First Name:  "+firstName+ ",  Last Name:  "+lastName+",  Email:  "+email+",  Phone:  "+phone);
+
+    public void printContact() {
+        System.out.println("First Name:  " + firstName + ",  Last Name:  " + lastName + ",  Email:  " + email + ",  Phone:  " + phone);
     }
 }
 
-class AddressBook{
+class AddressBook {
     private ArrayList<Contact> contacts;
-    AddressBook(){
-        contacts=new ArrayList<Contact>();
+
+    AddressBook() {
+        contacts = new ArrayList<Contact>();
     }
-    public void addContact(String firstname, String lastname, String email, String phone){
+
+    public void addContact(String firstname, String lastname, String email, String phone) {
         contacts.add(new Contact(firstname, lastname, email, phone));
     }
+
     public void printAddressBook() {
-        int indexOfContact=1;
+        int indexOfContact = 1;
         for (Contact singleContact : contacts) {
             System.out.print((indexOfContact++) + ": ");
             singleContact.printContact();
         }
     }
-    public void searchContact(int criteria,String searchQuery){
-        switch(criteria){
-            case 1:
-                for (int i=0; i<contacts.size(); i++){
-                    if(contacts.get(i).getFirstName().contains(searchQuery)) (contacts.get(i)).printContact();
+
+    public void searchContact(String criteria, String searchQuery) {
+        switch (criteria) {
+            case "1":
+                for (int i = 0; i < contacts.size(); i++) {
+                    if (contacts.get(i).getFirstName().contains(searchQuery)) (contacts.get(i)).printContact();
                     else System.out.println("\nThat contact doesn't exist");
-                } break;
-            case 2:
-                for (int i=0; i<contacts.size(); i++){
-                    if(contacts.get(i).getLastName().contains(searchQuery)) (contacts.get(i)).printContact();
+                }
+                break;
+            case "2":
+                for (int i = 0; i < contacts.size(); i++) {
+                    if (contacts.get(i).getLastName().contains(searchQuery)) (contacts.get(i)).printContact();
                     else System.out.println("\nThat contact doesn't exist");
-                } break;
-            case 3:
-                for (int i=0; i<contacts.size(); i++){
-                    if(contacts.get(i).getEmail().contains(searchQuery)) (contacts.get(i)).printContact();
+                }
+                break;
+            case "3":
+                for (int i = 0; i < contacts.size(); i++) {
+                    if (contacts.get(i).getEmail().contains(searchQuery)) (contacts.get(i)).printContact();
                     else System.out.println("\nThat contact doesn't exist");
-                } break;
-            case 4:
-                for (int i=0; i<contacts.size(); i++){
-                    if(contacts.get(i).getPhone().contains(searchQuery)) (contacts.get(i)).printContact();
+                }
+                break;
+            case "4":
+                for (int i = 0; i < contacts.size(); i++) {
+                    if (contacts.get(i).getPhone().contains(searchQuery)) (contacts.get(i)).printContact();
                     else System.out.println("\nThat contact doesn't exist");
-                } break;
+                }
+                break;
             default:
                 System.out.println("\n That contact doesn't exist");
         }
     }
-    public void deleteContact(int deleteContactChoice){
-        if(deleteContactChoice-1 < contacts.size() && deleteContactChoice>0) {
+
+    public void deleteContact(int deleteContactChoice) {
+        if (deleteContactChoice - 1 < contacts.size() && deleteContactChoice > 0) {
             contacts.remove(deleteContactChoice - 1);
         } else {
             System.out.println("\nNo such contact exists");
         }
     }
-    public void modifyContact(int modifyContactChoice, int attrToBeModified, String newValue){
-        switch(attrToBeModified){
-            case 1:
-                contacts.get(modifyContactChoice-1).setFirstName(newValue);
-                System.out.println("First name successfully changed to "+ newValue + "!\n");
+
+    public void modifyContact(int modifyContactChoice, String attrToBeModified, String newValue) {
+        switch (attrToBeModified) {
+            case "1":
+                contacts.get(modifyContactChoice - 1).setFirstName(newValue);
+                System.out.println("First name successfully changed to " + newValue + "!\n");
                 break;
-            case 2:
-                contacts.get(modifyContactChoice-1).setLastName(newValue);
-                System.out.println("Last name successfully changed to "+newValue + "!\n");
+            case "2":
+                contacts.get(modifyContactChoice - 1).setLastName(newValue);
+                System.out.println("Last name successfully changed to " + newValue + "!\n");
                 break;
-            case 3:
-                contacts.get(modifyContactChoice-1).setEmail(newValue);
-                System.out.println("Email address successfully changed to "+ newValue + "!\n");
+            case "3":
+                contacts.get(modifyContactChoice - 1).setEmail(newValue);
+                System.out.println("Email address successfully changed to " + newValue + "!\n");
                 break;
-            case 4:
-                contacts.get(modifyContactChoice-1).setPhone(newValue);
+            case "4":
+                contacts.get(modifyContactChoice - 1).setPhone(newValue);
                 System.out.println("Phone number successfully changed to " + newValue + "!\n");
                 break;
             default:
@@ -112,12 +132,14 @@ class AddressBook{
                 break;
         }
     }
-    public boolean validateContact(int modifyContactChoice){
-        if((modifyContactChoice-1 < contacts.size())&& (modifyContactChoice)>0) return true;
+
+    public boolean validateContact(int modifyContactChoice) {
+        if ((modifyContactChoice - 1 < contacts.size()) && (modifyContactChoice) > 0) return true;
         else return false;
     }
-    public boolean isEmpty(){
-        if(contacts.size()==0) return true;
+
+    public boolean isEmpty() {
+        if (contacts.size() == 0) return true;
         else return false;
     }
 }
@@ -143,27 +165,17 @@ class AddressBookProject {
                     break;
 
                 case "2":
-                    if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
-                    else {
-                        searchingContact(scanner, addr);
-                    }
+                    searchingContact(scanner, addr);
                     break;
 
                 case "3":
-                    if (addr.isEmpty()) {
-                        System.out.println("The Address Book is currently empty");
-                    } else {
-                        addr.printAddressBook();
-                        modifyingContact(scanner, addr);
-                    }
+                    addr.printAddressBook();
+                    modifyingContact(scanner, addr);
                     break;
 
                 case "4":
-                    if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
-                    else {
-                        addr.printAddressBook();
-                        removingContact(scanner,addr);
-                    }
+                    addr.printAddressBook();
+                    removingContact(scanner, addr);
                     break;
                 case "5":
                     if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
@@ -174,7 +186,7 @@ class AddressBookProject {
                 default:
                     if (!userInput.equals("e")) System.out.println("Invalid value please try again");
             }
-        } while (!userInput.equals("e"));
+        } while (true);
     }
 
     public static void addingContact(Scanner scanner, AddressBook addr) {
@@ -190,48 +202,56 @@ class AddressBookProject {
     }
 
     public static void searchingContact(Scanner scanner, AddressBook addr) {
-        System.out.println("\nPlease choose the criteria of your search: 1-By First Name, 2-By Last Name, 3-By Email, 4-By Phone");
-        int searchCriteria = scanner.nextInt();
-        scanner.nextLine();
-        if (searchCriteria < 1 || searchCriteria > 4)
-            System.out.println("Invalid criteria, returning to main menu");
+        if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
         else {
-            System.out.println("Add the keyword you would like to search");
-            String searchQuery = scanner.nextLine();
-            addr.searchContact(searchCriteria, searchQuery);
+            System.out.println("\nPlease choose the criteria of your search: 1-By First Name, 2-By Last Name, 3-By Email, 4-By Phone");
+            String searchCriteria = scanner.nextLine();
+            if (searchCriteria.compareTo("1") < 0 || searchCriteria.compareTo("4") > 0)
+                System.out.println("Invalid criteria, returning to main menu");
+            else {
+                System.out.println("Add the keyword you would like to search");
+                String searchQuery = scanner.nextLine();
+                addr.searchContact(searchCriteria, searchQuery);
+            }
         }
     }
 
     public static void modifyingContact(Scanner scanner, AddressBook addr) {
-        System.out.println("\nPlease choose the number of the contact that you would like to edit");
-        int modifyContactChoice = scanner.nextInt();
-        scanner.nextLine();
-        if (addr.validateContact(modifyContactChoice)) {
-            int attrToBeModified;
-            boolean flag = true;
-            do {
-                System.out.println("Please choose the attribute to be modified: (1-First Name, 2-Last Name, 3-Email, 4-Phone, 0-cancel)");
-                attrToBeModified = scanner.nextInt();
-                scanner.nextLine();
-                if (attrToBeModified < 0 || attrToBeModified > 4) {
-                    System.out.println("That attribute doesn't exist, please choose again");
-                } else if (attrToBeModified != 0) {
-                    System.out.println("Please enter the new value");
-                    String newValue = scanner.nextLine();
-                    addr.modifyContact(modifyContactChoice, attrToBeModified, newValue);
-                    System.out.println("Would you like to make additional changes? y/n");
-                    String additionalChange = scanner.nextLine();
-                    if (additionalChange.equals("n")) {
-                        flag = false;
+        if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
+        else {
+            System.out.println("\nPlease choose the number of the contact that you would like to edit");
+            int modifyContactChoice = scanner.nextInt();
+            scanner.nextLine();
+            if (addr.validateContact(modifyContactChoice)) {
+                String attrToBeModified;
+                boolean flag = true;
+                do {
+                    System.out.println("Please choose the attribute to be modified: (1-First Name, 2-Last Name, 3-Email, 4-Phone, 0-cancel)");
+                    attrToBeModified = scanner.nextLine();
+                    if ((attrToBeModified.compareTo("0") < 0) || (attrToBeModified.compareTo("4") > 0)) {
+                        System.out.println("That attribute doesn't exist, please choose again\n");
+                    } else if (attrToBeModified != "0") {
+                        System.out.println("Please enter the new value");
+                        String newValue = scanner.nextLine();
+                        addr.modifyContact(modifyContactChoice, attrToBeModified, newValue);
+                        System.out.println("Would you like to make additional changes? y/n");
+                        String additionalChange = scanner.nextLine();
+                        if (additionalChange.equals("n")) {
+                            flag = false;
+                        }
                     }
-                }
-            } while (attrToBeModified != 0 && flag);
-        } else System.out.println("\nNo such contact exists");
+                } while (attrToBeModified != "0" && flag);
+            } else System.out.println("\nNo such contact exists");
+        }
     }
-    public static void removingContact(Scanner scanner, AddressBook addr){
-        System.out.println("\nPlease choose the number of the contact that you would like to delete");
-        int deleteContactChoice = scanner.nextInt();
-        scanner.nextLine();
-        addr.deleteContact(deleteContactChoice);
+
+    public static void removingContact(Scanner scanner, AddressBook addr) {
+        if (addr.isEmpty()) System.out.println("The Address Book is currently empty");
+        else {
+            System.out.println("\nPlease choose the number of the contact that you would like to delete");
+            int deleteContactChoice = scanner.nextInt();
+            scanner.nextLine();
+            addr.deleteContact(deleteContactChoice);
+        }
     }
 }
